Alerting: Split Loki alert state history into reader and writer.
The historian app will provide the API to read alert state history, though it
does not (and will not) have the capability to write history. This refactor
makes it possible to create a reader without a writer.
